% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brcancer}
\alias{brcancer}
\title{German Breast Cancer Study Data}
\format{
A data frame with 686 rows and 14 variables:
\itemize{
\item \code{id} A numeric vector;
\item \code{hormon} Hormonal therapy;
\item \code{x1} Age, in years;
\item \code{x2} Menopausal status;
\item \code{x3} Tumour size, mm;
\item \code{x4} Tumour grade;
\item \code{x5} Number of positive nodes;
\item \code{x6} Progesterone receptor, fmol;
\item \code{rectime} Recurrence-free survival time, days;
\item \code{censrec} Censoring indicator;
\item \code{x4a} Tumour grade >=2;
\item \code{x4b} Tumour grade == 3;
\item \code{x5e} exp(-0.12 * x5).
}
}
\usage{
brcancer
}
\description{
German breast cancer study data, imported from Stata 16.
}
\examples{
data("brcancer", package = "KMunicate")
}
\references{
http://www.stata-press.com/data/r16/brcancer.dta
}
\keyword{datasets}
