% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cancer2}
\alias{cancer2}
\title{Patient Survival in Drug Trial}
\format{
A data frame with 48 rows and 4 variables:
\itemize{
\item \code{studytime} Months to death or end of follow-up;
\item \code{died} Event indicator variable, \code{died = 1} if a patient died;
\item \code{drug} Drug type, with \code{drug = 1} being placebo;
\item \code{age} Age of a patient at baseline.
}
}
\usage{
cancer2
}
\description{
Patient survival in drug trial data, imported from Stata 16.
}
\note{
The dataset is named \code{cancer2} to avoid name collision with the \code{cancer} dataset from the survival package.
}
\examples{
data("cancer2", package = "KMunicate")
}
\references{
http://www.stata-press.com/data/r16/cancer.dta
}
\keyword{datasets}
