% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMunicate-package.R, R/KMunicate.R
\docType{package}
\name{KMunicate}
\alias{KMunicate}
\title{KMunicate-Style Kaplan–Meier Plots}
\usage{
KMunicate(
  fit,
  time_scale,
  .theme = ggplot2::theme_minimal(),
  .color_scale = NULL,
  .fill_scale = NULL,
  .xlab = "Time",
  .alpha = 0.25,
  .rel_heights = NULL
)
}
\arguments{
\item{fit}{A \code{survfit} object.}

\item{time_scale}{The time scale that will be used for the x-axis and for the summary tables.}

\item{.theme}{\code{ggplot} theme used by the plot. Defaults to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.}

\item{.color_scale}{Colour scale used for the plot. Has to be a \verb{scale_colour_*} component, and defaults to \code{NULL} where the default colour scales will be used.}

\item{.fill_scale}{Fill scale used for the plot. Has to be a \verb{scale_fill_*} component, and defaults to \code{NULL} where the default fill scales will be used.}

\item{.xlab}{Label for the horizontal axis, defaults to \emph{Time}.}

\item{.alpha}{Transparency of the point-wise confidence intervals}

\item{.rel_heights}{Override default relative heights of plots and tables. See \code{\link[cowplot:plot_grid]{cowplot::plot_grid()}} for more details on how to use this argument.}
}
\value{
A KMunicate-style \code{ggplot} object.
}
\description{
Produce Kaplan–Meier plots in the style recommended following the KMunicate study \url{doi:10.1136/bmjopen-2019-030215}.

Produce Kaplan–Meier plots in the style recommended following the KMunicate study \url{doi:10.1136/bmjopen-2019-030215}.
}
\examples{
library(survival)
data("cancer", package = "KMunicate")
KM <- survfit(Surv(studytime, died) ~ drug, data = cancer)
time_scale <- seq(0, max(cancer$studytime), by = 7)
KMunicate(fit = KM, time_scale = time_scale)
}
\author{
Alessandro Gasparini (alessandro.gasparini@ki.se)
}
