% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMunicate-package.R
\docType{package}
\name{KMunicate_package}
\alias{KMunicate_package}
\alias{KMunicate-package}
\title{KMunicate-Style Kaplan–Meier Plots}
\description{
Produce Kaplan–Meier plots in the style recommended following the KMunicate study (\doi{10.1136/bmjopen-2019-030215}).
}
\author{
Alessandro Gasparini (alessandro@ellessenne.xyz)
}
